requirements.txt
qwen-agent>=0.0.11
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
